Add basic PlacesFetcher
Usage of the PlacesFetcher is simple at this point.
Instantiate an instance of it, set properties for latitude,
longitude, callback object, etc. Once fetcher has received
a response for all requests (multiple placeTypes can be fetched),
it lets the callback object know.

NOTE: As of now, if a request for a placeType fails, no handling is
done as of now.